入力を3次元配列に
