+++ app/alpha_zero/utils.py
@@ -1,5 +1,9 @@
+from typing import List
+
+import numpy as np
+
 from ..games.arena import Arena
-from ..games.players import Player
+from ..games.player_base import Player
 from ..games.game import Game
 
 
@@ -10,3 +14,8 @@ def eval_player(
     arena2 = Arena(standard_player, eval_player, game)
     r = (arena1.play_games(num_game) - arena2.play_games(num_game)) / 2
     return r
+
+
+def get_board_view(cboard: List[List[float]]):
+    board = np.array(cboard)
+    return [(board == 1).tolist(), (board == -1).tolist()]
