+++ app/games/player_base.py
@@ -1,0 +1,12 @@
+from typing import List
+
+
+class Player:
+    def __init__(self):
+        pass
+
+    def play(self, board: List[List[float]]) -> int:
+        raise NotImplementedError()
+
+    def reset(self) -> None:
+        return
